requirements.txt
contourpy==1.3.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
easyocr==1.7.2
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
imageio==2.36.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.1
ninja==1.11.1.1
numpy==2.1.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==11.0.0
pyclipper==1.3.0.post5
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sympy==1.13.3
tifffile==2024.9.20
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
